Support for C_DeriveKey and minimal support for Secret Key Objects
When using ECDH for encryption, a secret key is derived from
the private key using C_DeriveKey. The result is then stored
in a secret key object CKO_SECRET_KEY. This can then be
retrieved by the caller by reading the CKA_VALUE of the secret
key object.

Thus Secret Key Object support is now required in OpenSC.
These changes are the minimal set needed to support this
use with the PIV card using CKM_ECDH1_COFACTOR_DERIVE.

Only PKCS#11 session secret key objects are supported.
Although the derivation is done on the card, the PIV returns
the result as part of the operation. Some cards may be
beable to store a derived key on the card, and use it
on the card. That would be a future project.

The only operations on the secret key object is to read
attributes and destroy the object.  No secret key mechanisums
are added. No support to use the secret key in any other way
is added.

The only key type supported is CKK_GENERIC_SECRET. Thus
the secret key object CVA_VALUE is just a string.

The only CK_ECDH1_DERIVE_PARAMS kdf is CKD_NULL.