minidriver: 'write' mode
initial implementation of the OpenSC minidriver in write mode:

- in 'write' and 'read-only' modes:
-- 'cardcf' content is created with the first successfull method in the following order:
--- if present the on-card pkcs#15 DATA object 'CSP':'cardcf' - use it's content;
--- if present 'lastUpdate' attribute of tokenInfo - use CRC-32 calculated on it's value;
--- random value.

-- 'cmapfile' (containers) is emulated from existing privateKey pkcs#15 objects.
   If present the on-card pkcs#15 DATA object 'CSP':'cmapfile' - it's content used to
   update the non-pkcs#15 attributes of containers.

- 'write' mode:
- 'write' mode is activated by setting to 'false' the 'md_read_only' option in
  the 'card_atr' section of OpenSC configuration file;
-- every 'WriteFile' on 'cardcf' updates the on-card pkcs#15 DATA object 'CSP':'cardcf'.
-- on 'DeAuthenticate' call the 'cmapfile' content is encoded and written into the
   on-card pkcs#15 DATA object 'CSP':'cmapfile'.

Tested with 'AMOS IAS/ECC' card in IE on Windows XP platform.
Test consisted in the decentrilized card enrollment, followed by the authentication to access the protected cite.

For the unknown reason, when generating key in IE, the BaseCSP tries firstly to import the 'soft' key,
instead of generating one on-card. (If minidriver refuse this attempt, BaseCSP generates key on-card).
This 'feature' gave the possibility to test key generation and key import.
No other application where used for tests.
Still needs to be tested on the other windows versions.