pkcs15: exported procedure 'get_application_by_type'
For the multi-applications card this procedure returns the first application of a given type.
Application type defined in the 'framework'/'pkcs15' section of opensc.conf.
Actually defined 'generic' and 'protected' types.
This procedure is to be used in minidriver and in 'pkcs15' framework
and replaces the internal pkcs11 static procedure.